+++ src/test/java/br/org/soujava/pomeditor/RollbackMojoTest.java
@@ -14,42 +14,34 @@
  * limitations under the License.
  */
 
-package io.github.arrudalabs.pomeditor;
+package br.org.soujava.pomeditor;
 
+import org.apache.maven.plugin.MojoExecutionException;
+import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.plugin.logging.Log;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Captor;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import java.nio.file.Path;
+import java.util.function.BiConsumer;
+
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
-class GreetingMojoTest {
+class RollbackMojoTest {
 
     @Mock
-    Log logger;
-
-    @Captor
-    ArgumentCaptor<String> output;
+    BiConsumer<Log, Path> rollbackFunction;
 
     @Test
-    void shouldExecuteWithNoErrors() {
-        GreetingMojo mojo = newMojo();
-        mojo.username = "Max";
+    void execute() throws MojoExecutionException, MojoFailureException {
+        var mojo = new RollbackMojo();
+        mojo.rollbackFunction = this.rollbackFunction;
         mojo.execute();
-        verify(logger, atLeastOnce()).info(output.capture());
-        assertEquals("Hi, Max!", output.getValue());
+        verify(rollbackFunction, atLeastOnce()).accept(any(Log.class), any(Path.class));
     }
-
-    private GreetingMojo newMojo() {
-        GreetingMojo mojo = new GreetingMojo();
-        mojo.setLog(logger);
-        return mojo;
-    }
-
 }
