+++ src/test/java/br/org/soujava/pomeditor/it/AddDependencyMojoIT.java
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-package io.github.arrudalabs.pomeditor.it;
+package br.org.soujava.pomeditor.it;
 
 import com.soebes.itf.jupiter.extension.MavenGoal;
 import com.soebes.itf.jupiter.extension.MavenJupiterExtension;
-import com.soebes.itf.jupiter.extension.MavenOption;
 import com.soebes.itf.jupiter.extension.MavenTest;
 import com.soebes.itf.jupiter.extension.SystemProperty;
 import com.soebes.itf.jupiter.maven.MavenExecutionResult;
