+++ src/test/resources-its/br/org/soujava/pomeditor/it/AddDependencyMojoIT/has_dependency_already/pom.xml
