+++ src/test/resources-its/br/org/soujava/pomeditor/it/AddDependencyMojoIT/has_dependency_but_with_version_lower/pom.xml
