+++ src/test/resources-its/br/org/soujava/pomeditor/it/AddDependencyMojoIT/no_dependencies_declarated/pom.xml
