+++ src/test/resources-its/br/org/soujava/pomeditor/it/AddDependencyMojoIT/with_other_dependencies/pom.xml
