+++ src/test/resources-its/br/org/soujava/pomeditor/it/GreetingMojoIT/hello/pom.xml
